Moved Mention controller to routes
